Add sign-on-methods-core component
